lambda-functions/rds-query/requirements.txt
pymysql==1.1.0
psycopg2-binary==2.9.9